:art: Add new Badges and put old badges to CardButtons
