+++ client/src/stories/Card.stories.js
@@ -2,6 +2,7 @@ import React from "react";
 import { Badge } from "../components/Badge";
 
 import { Card } from "../components/Card";
+import { CardButton } from "../components/CardButton";
 
 export default {
   title: "DispoDisco/Cards",
@@ -9,6 +10,7 @@ export default {
 };
 
 const Template = (args) => <Card {...args} />;
+const CardButtonTemplate = (args) => <CardButton {...args} />;
 
 export const NormalRide = Template.bind();
 NormalRide.args = {
@@ -56,4 +58,19 @@ OnTimeRide.args = {
 export const ConcurrentRide = Template.bind();
 ConcurrentRide.args = {
   type: "concurrentRide",
+};
+
+export const Rider = CardButtonTemplate.bind(0);
+Rider.args = {
+  type: "rider",
+  label: "🚴‍♀️ Elena",
+};
+export const Timer = CardButtonTemplate.bind(0);
+Timer.args = {
+  type: "timer",
+  label: "1:30h",
+};
+export const Info = CardButtonTemplate.bind(0);
+Info.args = {
+  type: "info",
 };
